Development 1.8 (#38)
* Addressed #35 - Switched to slf4j

* Made `AnnotatedCommandCompiler` an interface and moved previously existing code to `com.jagrosh.jdautilities.commandclient.impl` under the name `AnnotatedCommandCompilerImpl`.
