+++ build.gradle
@@ -25,8 +25,8 @@ plugins {
     id 'com.github.johnrengelman.shadow' version '2.0.1'
 }
 
-def versionObj = new Version(major: 1, minor: 7)
-def jdaVersion = '3.3.0_265'
+def versionObj = new Version(major: 1, minor: 8)
+def jdaVersion = '3.3.1_284'
 
 group = 'com.jagrosh'
 archivesBaseName = project.name
@@ -149,6 +149,11 @@ build {
     dependsOn javadocJar
     dependsOn sourcesJar
     dependsOn shadowJar
+
+    jar.mustRunAfter clean
+    javadocJar.mustRunAfter jar
+    sourcesJar.mustRunAfter javadocJar
+    shadowJar.mustRunAfter sourcesJar
 }
 
 bintrayUpload {
