+++ src/main/java/com/jagrosh/jdautilities/JDAUtilitiesInfo.java
@@ -21,7 +21,7 @@ package com.jagrosh.jdautilities;
  */
 public class JDAUtilitiesInfo {
     public static final String VERSION_MAJOR = "1";
-    public static final String VERSION_MINOR = "7";
+    public static final String VERSION_MINOR = "8";
     public static final String VERSION = VERSION_MAJOR+"."+VERSION_MINOR;
     public static final String GITHUB = "https://github.com/JDA-Applications/JDA-Utilities";
     public static final String AUTHOR = "JDA-Applications";
