+++ src/main/java/com/jagrosh/jdautilities/commandclient/package-info.java
@@ -15,27 +15,32 @@
  */
 
 /**
- * <h1><b>CommandClient Package</b></h1>
- * 
  * <p>Items in this package pertain to the {@link com.jagrosh.jdautilities.commandclient.CommandClient CommandClient} and
  * {@link com.jagrosh.jdautilities.commandclient.Command Commands}.
  * 
  * <p>All of the contents are used heavily in the {@link com.jagrosh.jdautilities.commandclient.impl.CommandClientImpl CommandClientImpl}, 
  * and are summarized as follows:
  * <ul>
+ *     <li>{@link com.jagrosh.jdautilities.commandclient.AnnotatedModuleCompiler AnnotatedModuleCompiler}
+ *     <br>An interface to create Commands from annotated objects (More info on annotated commands can be found in the
+ *     {@link com.jagrosh.jdautilities.commandclient.annotation.JDACommand JDACommand} documentation).</li>
+ *     <br>
+ *     <li>{@link com.jagrosh.jdautilities.commandclient.CommandBuilder CommandBuilder}
+ *     <br>An chain builder for Commands.</li>
+ *     <br>
  *     <li>{@link com.jagrosh.jdautilities.commandclient.Command Command}
  *     <br>An abstract class that can be inherited by classes to create Commands compatible with the {@code CommandClientImpl}.</li>
- *     
+ *     <br>
  *     <li>{@link com.jagrosh.jdautilities.commandclient.CommandClient CommandClient}
  *     <br>An interface used for getting info set when building a {@code CommandClientImpl}.</li>
- *     
+ *     <br>
  *     <li>{@link com.jagrosh.jdautilities.commandclient.CommandClientBuilder CommandClientBuilder}
  *     <br>A builder system used to create a {@code CommandClientImpl} across several optional chained methods.</li>
- *     
+ *     <br>
  *     <li>{@link com.jagrosh.jdautilities.commandclient.CommandEvent CommandEvent}
  *     <br>A wrapper for a {@link net.dv8tion.jda.core.events.message.MessageReceivedEvent MessageReceivedEvent}, {@code CommandClient}, and String arguments. 
  *     The main basis for carrying information to be used in Commands.</li>
- *     
+ *     <br>
  *     <li>{@link com.jagrosh.jdautilities.commandclient.CommandListener CommandListener}
  *     <br>An interface to be provided to a {@code CommandClientImpl} that can provide Command operations depending on the outcome of the call.</li>
  * </ul>
